[NuGet] Allow forcing a NuGet package uninstall.
If a NuGet package is not restored and an uninstall of that NuGet
package is attempted then a restore will be attempted. If the restore
succeeds then the NuGet package will be uninstalled. If the restore
fails then the user will be prompted asking if they want to force the
NuGet package to be removed. Forcing a NuGet package to be removed
will look for any references and import associated with the NuGet
package based on the packages install path. It will then remove the
NuGet package from the packages.config file. This feature is only
available if the project uses a packages.config file. A project that
uses a project.json file will not fail to uninstall since removing
the NuGet package will only modify the project.json file.

+++ main/src/addins/MonoDevelop.PackageManagement/MonoDevelop.PackageManagement.Commands/PackageReferenceNodeCommandHandler.cs
@@ -51,22 +51,23 @@ namespace MonoDevelop.PackageManagement.Commands
 
 		void RemovePackage (PackageReferenceNode packageReferenceNode, ProgressMonitorStatusMessage progressMessage)
 		{
-			var solutionManager = PackageManagementServices.Workspace.GetSolutionManager (packageReferenceNode.Project.ParentSolution);
-			var action = new UninstallNuGetPackageAction (solutionManager, packageReferenceNode.Project) {
-				PackageId = packageReferenceNode.Id
-			};
-			// TODO - handle unrestored NuGet packages.
-			//if (action.Package != null) {
-				PackageManagementServices.BackgroundPackageActionRunner.Run (progressMessage, action);
-			//} else {
-			//	ShowMissingPackageError (progressMessage, packageReferenceNode);
-			//}
+			IPackageAction action = CreateUninstallPackageAction (packageReferenceNode);
+			PackageManagementServices.BackgroundPackageActionRunner.Run (progressMessage, action);
 		}
 
-		void ShowMissingPackageError (ProgressMonitorStatusMessage progressMessage, PackageReferenceNode packageReferenceNode)
+		IPackageAction CreateUninstallPackageAction (PackageReferenceNode packageReferenceNode)
 		{
-			string message = GettextCatalog.GetString ("Unable to find package {0} {1} to remove it from the project. Please restore the package first.", packageReferenceNode.Id, packageReferenceNode.Version);
-			PackageManagementServices.BackgroundPackageActionRunner.ShowError (progressMessage, message);
+			var solutionManager = PackageManagementServices.Workspace.GetSolutionManager (packageReferenceNode.Project.ParentSolution);
+			if (packageReferenceNode.NeedsRestoreBeforeUninstall ()) {
+				return new RestoreAndUninstallNuGetPackageAction (solutionManager, packageReferenceNode.Project) {
+					PackageId = packageReferenceNode.Id,
+					Version = packageReferenceNode.Version
+				};
+			}
+
+			return new UninstallNuGetPackageAction (solutionManager, packageReferenceNode.Project) {
+				PackageId = packageReferenceNode.Id
+			};
 		}
 
 		[CommandUpdateHandler (EditCommands.Delete)]
